% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ats_fixrate.R
\name{ats_fixrate}
\alias{ats_fixrate}
\title{Change fix rate on ATS collars}
\usage{
ats_fixrate(user, pass, fixrate)
}
\arguments{
\item{user}{ATS IDAQ username (character vector)}

\item{pass}{ATS IDAQ password (character vector)}

\item{fixrate}{fixrate in hours, one of: (4,6,8,12,24,48,72,96,120,144,168)}
}
\description{
Automates changing fix rates on all collars in an ATS account
}
\examples{
\donttest{ats_fixrate(user='username', pass='password', fixrate = 48)}
}
\keyword{ATS,}
\keyword{fix}
\keyword{fixrate}
\keyword{rate,}
