% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGBFun.R
\name{BGBFun}
\alias{BGBFun}
\title{BiVariate Gaussian Bridge Stats}
\usage{
BGBFun(data, xname, yname, timename, idname, projstring, ncpus, msize = 21,
  winsize = 43)
}
\arguments{
\item{data}{GPS data either as data.frame}

\item{xname}{name of X coordinate column}

\item{yname}{name of Y coordinate column}

\item{timename}{name of POSIX time/date column}

\item{idname}{name of unique identifier for each animal column}

\item{projstring}{Correct proj4string for the data/coordinates}

\item{ncpus}{Number of CPUs for parallel processing. Recommend 1-2 less than max.}

\item{msize}{Margin size for dynBGB. Defaults to 21.}

\item{winsize}{Window size for dynBGB. Defaults to 43.}
}
\value{
Resulting object is a data.frame of original data with dynBGB parameters added as columns. See Kraunstauber
2015 for more info.

\item{paraSd}{Parallel Variance Measure}
\item{orthSd}{orthogonal variance measure}
}
\description{
Create BiVariate Gaussian Bridge Stats using the move package. Specifically
creates parallel and orthogonal variance measures in format to append to data.frame of movement data.
This can be a lengthy process and therefore, this code is internally parallelized. You must give an
ncpus > 1 to run in parallel.
}
\examples{
\donttest{BGBFun(data=df, xname='Easting',yname='Northing',timename='TelemDate',idname='Serial',
projstring='+proj=longlat +datum=WGS84 +no_defs',ncpus=6,msize=21,winsize=43)}

}
\keyword{bgb}
\keyword{bivariate}
\keyword{bridge,}
\keyword{dynbgb,}
\keyword{gaussian}
