% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalfMark.R
\name{CalfMark}
\alias{CalfMark}
\title{Daily Calf Update}
\usage{
CalfMark(ATSUser, ATSPass, tempdir, ncpu, lookup, from = NA, to = NA,
  subject = NA, SP = NA, progpath = NA, username = NA,
  password = NA, email = "no", systype = "Win")
}
\arguments{
\item{ATSPass}{character vector of ATS passwords}

\item{tempdir}{temporary folder for downloading data}

\item{ncpu}{number of CPU cores for multithreaded work}

\item{lookup}{path to lookup table}

\item{vecpath}{path to vectronic data folder}

\item{ATSUsers}{character vector of ATS user names}

\item{spp}{Species of animal}
}
\value{
Creates all data needed for fawnmark
}
\description{
Run daily parturition calf updates data manipulation and plotting
}
\examples{
\donttest{vecVit<-vecVitDat(path='F:/Box Sync/DEER/Data/Vectronic/VecData')}

}
\keyword{fawnmark,}
\keyword{prep}
