% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capdat.R
\name{capdat}
\alias{capdat}
\title{Conversion of GPS data to GPX/KML for use in capture events.}
\usage{
capdat(data, uid, keepvec = NULL, folder, TD, proj, spp = "animal")
}
\arguments{
\item{data}{data.frame of relocation data. Should contain unique ID field, POSIXct time/date,
x coordinate (column name must be 'x') and y coordinate (column name must be 'y').}

\item{uid}{Unique animal identifier column header, frequency or serial.
This is what will show up on GPS/KML. Likely want this to be column containing frequency.}

\item{keepvec}{(OPTIONAL) Character vector of animals to keep}

\item{folder}{Location to save gpx/kml file to. (eg. 'C:/Users/mhayes1/Desktop/')}

\item{TD}{Column name for POSIX time/date field.}

\item{proj}{proj4string of Easting/Northing values. Can be obtained by running proj4string(Irid[[1]])
on output of Irid function or running proj4string on the SpatialPointsDataFrame.}

\item{spp}{(OPTIONAL) This is the layer name of the kml as viewed in Google Earth}
}
\value{
Saves a .gpx, .kml and .csv file of most recent relocation for each animal in 'data'. Can be used
directly with output of GStar or Irid after converting them to data.frames.
}
\description{
Convert your spatial data into GPX and KML files for use in captures. Noteably, used
for helicopter recaptures of previously marked animals. Takes input data, finds last fix for each animal
and converts this data to a GPX and KML file. Also saves a csv with the same data so user can inspect
the data and ensure that last fix occured recently. Helpful to determine dead or failed collars.
}
\examples{
\donttest{ark<-GStar(username=username,password=password)
a<-ark[[1]]
proj<-proj4string(a)
a<-as.data.frame(a)
names(a)[3:4]<-c('y','x')
capdat(data=a,uid='CollarSerialNumber',fold='C:/Users/mhayes1/Desktop/test/',
      TD='TelemDate',proj=proj,spp='Deers')}

}
\keyword{capture,}
\keyword{gpx,}
\keyword{kml}
