% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanFun.R
\name{cleanFun}
\alias{cleanFun}
\title{GPS Data Cleaning}
\usage{
cleanFun(data, fixstat = "X2D.3D", hdopC = "HDOP", cval = 3, hval = 10,
  filename, type = "irid", spp = "Elk")
}
\arguments{
\item{data}{SpatialPointsDataFrame as outputed from GStar or Irid. May work with others that are in UTM 12N.}

\item{fixstat}{name of 2d/3d fix column. Defaults to 'X2D.3D'}

\item{hdopC}{name of HDOP column. Defaults to 'HDOP'}

\item{cval}{Cutoff for 2d/3d fix status. Valid values are 2 or 3. 3 will only use 3d fixes while 2 uses 2 and 3.}

\item{hval}{Cutoff for hdop quantile cutoff. Default is 10 which is equalt to 90th quantile.}

\item{filename}{Full path to store cleaning report file.}

\item{type}{Character vector of data type, either: 'gstar' or 'irid'}

\item{spp}{Character vector of species name (eg. 'Deer')}
}
\value{
Resulting object is a list of three elements. First element is a SpatialPointsDataFrame of
all the GPS data which met the cleaning critera, the second element is all the "bad" data cleaned by
set parameters and a report of what data was cleaned and for what reason.

Accessing the "good" spatial data is done by cleanFun()[[1]]
Accessing the "bad" spatial data is done by cleanFun()[[1]]
Accessing the cleaning report is done by cleanFun()[[1]]
}
\description{
Cleans output from GStar or Irid. Cleaning is user-defined and
based on either hdop or 2d/3d fix status. Outputs three results in list format: Cleaned Data,
"Bad" data and a report. Data is also cleaned if the 'dist' (from as.ltraj) is greater than the 98 percentile.
}
\examples{
\donttest{cleanFun(data, filename='C:/users/mhayes1/Desktop/Report',cval=3,hval=10,type='irid',spp='Elk')}

}
\keyword{clean,}
\keyword{cleaning}
\keyword{data}
