% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColDownload.R
\name{ColDownload}
\alias{ColDownload}
\title{Internet Collar Download}
\usage{
ColDownload(username = "", password = "", dirdown = "",
  cType = "ATS/IRID")
}
\arguments{
\item{username}{Username for Iridium Account (not stored) as character vector}

\item{password}{Password for Iridium Account (not stored) as character vector}

\item{dirdown}{Directory to download the file to}

\item{cType}{Type of collar/system (currently only supports 'ATS/IRID' and 'ATS/GSTAR')}
}
\value{
Resulting object is a list of Two elements. First element is a SpatialPointsDataFrame of
all the GPS data on the account, the second is a dataframe of all transmissions received by
the satellite

\strong{Spatial Data Column Description:}
\item{CollarSerialNumber}{ATS Designated Collar Serial Number}
\item{TelemDate}{POSIX Field of collare fixes in the USER'S timezone. NOTE: may need to alter timezone}
\item{HDOP}{Horizontal Dilution of Precision}
\item{NumSats}{Number of satellites used for GPS Fix}
\item{2D/3D}{Whether fix is a 2d or 3d fix. Values are either 2 or 3}
}
\description{
This function downloads Iridium GPS data from ATS servers. Internet connection
must be available and ATS servers online. User must have valid ATS Iridium account. In addition
mortality/battery warnings and transmissions are also downloaded and returned.
}
\examples{
\donttest{ColDownload(username='yourusername',password='yourpassword',dirdown='/home/mhayes1/Desktop/Testing/',cType='ATS/IRID')}

}
\keyword{Iridium,}
\keyword{iridium}
