% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombDat.R
\name{CombDat}
\alias{CombDat}
\title{Combine ATS and Vectronic Data}
\usage{
CombDat(vecpath, ATSUsers, ATSPass, tempdir)
}
\arguments{
\item{vecpath}{Path your Vectronic GPS Plus X Program saves CSV files to}

\item{ATSUsers}{a character vector of all ATS Iridium login Usernames}

\item{ATSPass}{a character vector of all ATS Iridium login passwords}

\item{tempdir}{a temporary directory to store files in when downloading from ATS site}
}
\value{
Resulting object is a spatial points data frame of your ATS and vectronics data
downloaded using GPS Plus X as a csv
\strong{Spatial Data Column Description:}
\item{CollarSerialNumber}{Collar Serial Number}
\item{TelemDate}{POSIX Field of collare fixes in the USER'S timezone. NOTE: may need to alter timezone}
\item{HDOP}{Horizontal Dilution of Precision}
\item{2D/3D}{Whether fix is a 2d or 3d fix. Values are either 2 or 3}
\item{Temperature}{Temperature reported by collar}
}
\description{
This function combines ATS Iridium data and the csv files downloaded by vectronic GPS Plus X
}
\examples{
\donttest{dd<-CombDat(vecpath='F:/Box Sync/DEER/Data/Vectronic/VecData',
ATSUsers=c('user1','user2'),
ATSPass = c('pass1','pass2'),
tempdir='C:/Users/mhayes1/Desktop/temp')}

}
\keyword{ats}
\keyword{vectronic,}
