% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElkRFPred.R
\name{ElkRFPred}
\alias{ElkRFPred}
\title{Elk Parturition Prediction with Machine Learning}
\usage{
ElkRFPred(jk)
}
\arguments{
\item{jk}{output of BGBFun}
}
\value{
data.frame of probability predictions for 0, 1, 2 (pre birth, neonate <48hrs old, neonate>48hrs old)
}
\description{
Use Machine Learning to predict parturition in elk
}
\examples{
\donttest{hg<-ElkRFPred(mdat2)}

}
\keyword{prediction}
