% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVec.R
\name{getVec}
\alias{getVec}
\title{Vectronic Collar Data Combine}
\usage{
getVec(vecpath)
}
\arguments{
\item{vecpath}{Path your Vectronic GPS Plus X Program saves CSV files to}
}
\value{
Resulting object is a spatial points data frame of your vectronics data
downloaded using GPS Plus X as a csv
\strong{Spatial Data Column Description:}
\item{CollarSerialNumber}{Collar Serial Number}
\item{TelemDate}{POSIX Field of collare fixes in the USER'S timezone. NOTE: may need to alter timezone}
\item{HDOP}{Horizontal Dilution of Precision}
\item{2D/3D}{Whether fix is a 2d or 3d fix. Values are either 2 or 3}
\item{Temperature}{Temperature reported by collar}
}
\description{
This function combines the csv files downloaded by vectronic GPS Plus X
}
\examples{
\donttest{getVec(vecpath='/home/mhayes1/Desktop/Testing/')}

}
\keyword{vectronic}
