% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLVarPrep.R
\name{MLVarPrep}
\alias{MLVarPrep}
\title{Prepare variables for Machine Learning Parturition prediction}
\usage{
MLVarPrep(jk)
}
\arguments{
\item{jk}{output of BGBFun}
}
\value{
data.frame of probability predictions for 0, 1, 2 (pre birth, neonate <12hrs old, neonate>12hrs&neonate<24hrs old)
}
\description{
Use Machine Learning to predict parturition
}
\examples{
\donttest{hg<-ElkRFPred(mdat2)}

}
\keyword{prediction}
