% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_BGBFun.R
\name{Part_BGBFun}
\alias{Part_BGBFun}
\title{Parturition BGB Creation}
\usage{
Part_BGBFun(data, xname, yname, timename, idname, projstring, ncpus)
}
\arguments{
\item{data}{"Cleaned" data}

\item{xname}{column name for x coordinates}

\item{yname}{column name for y coordinates}

\item{timename}{column name for posix time}

\item{idname}{column name for unique individual identifier}

\item{projstring}{character proj4string}

\item{ncpus}{Number of cpus to use}
}
\value{
Resulting object is the data with BGB variables attached
}
\description{
Creates BGB variables for predicting parturition timing in ungulates given longitudinal
relocation data via parallel processing. Internally used in parturition methodology.
}
\examples{
\donttest{dyn_dat <- BGBFun(data = dat, xname = "x", yname = "y", timename = "TelemDate", idname = "UAID", projstring = projection, ncpus = ncpus)}

}
\keyword{part,}
\keyword{parturition}
