% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_bindfun.R
\name{Part_bindfun}
\alias{Part_bindfun}
\title{Parturition Trajectory Binding}
\usage{
Part_bindfun(ti, outfun)
}
\arguments{
\item{ti}{trajectory data}

\item{outfun}{original data}
}
\value{
Binds together trajectory and original data
}
\description{
Binds trajectory variables with original data for ungulate partition prediction.
Internally used in parturition methodology.
}
\keyword{part,}
\keyword{parturition}
