% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_calcmoveStats.R
\name{Part_calcmoveStats}
\alias{Part_calcmoveStats}
\title{Wrapper function to create all variables necessary to create ungulate parturition
prediction models.}
\usage{
Part_calcmoveStats(dat, projstring, dataset, time.zone, folder,
  ncpus = ncpus, saveby = "Study", type = "Norm", imp = FALSE)
}
\arguments{
\item{dat}{data}

\item{projstring}{Experimental, leave as default (FALSE)}

\item{dataset}{Short name of dataset without space ('SMElk')}

\item{time.zone}{POSIX format timezone}

\item{folder}{character folder path ('/home/user/folder')}

\item{ncpus}{Number of cpus to use for parallel processing (minimum 2)}

\item{saveby}{Character vector, either "Study" to save a single file of all data or
"Animal" to save one file for each individual in your dataset}

\item{type}{Character vector, either 'Norm' for normal running or 'NPred' for new predictions.}
}
\value{
Original data with rolling candidate variables
}
\description{
Wrapper function to create all variables necessary to create ungulate parturition
prediction models.
}
\keyword{part,}
\keyword{parturition}
