% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_caretModPrep.R
\name{Part_caretModPrep}
\alias{Part_caretModPrep}
\title{Wrapper function to create all variables necessary to create ungulate parturition
prediction models.}
\usage{
Part_caretModPrep(folder, datatype, mean_date, bday_dat, idname = "UAID",
  sampsize = 80, ncpus)
}
\arguments{
\item{folder}{folder where ready data is stored '/home/user/Desktop/Partmod'}

\item{datatype}{Either 'Animal' if files for each animal or 'Study' if one file saved}

\item{mean_date}{julian day (as.numeric) of the mean date of parturition for your animal/area}

\item{bday_dat}{data.frame with two columns, UAID (character) and Date.of.Birth (as.Date)}

\item{idname}{character vector name of UAID column}

\item{sampsize}{percent of dataset to use to train model}

\item{nrun}{folder where ready data is stored '/home/user/Desktop/Partmod'}
}
\value{
Original data with rolling candidate variables
}
\description{
Wrapper function to create all variables necessary to create ungulate parturition
prediction models.
}
\keyword{part,}
\keyword{parturition}
