% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_fixrate.R
\name{Part_fixrate}
\alias{Part_fixrate}
\title{Rarefy Parturition Model}
\usage{

  Part_fixrate(input = "/home/puma/Desktop/RunningPart/Data/StarkeyElk_ReadyData.RData",
  name, saveby = "Study", ncpu = 75, basefolder, sampsize = 80,
  prepboots = 75, finboots = 100, max_fix = 12)
}
\arguments{
\item{input}{Username for Iridium Account (not stored) as character vector}

\item{name}{Password for Iridium Account (not stored) as character vector}

\item{saveby}{Directory to download the file to}

\item{ncpu}{Type of collar/system (currently only supports 'ATS/IRID' and 'ATS/GSTAR')}

\item{basefolder}{folder to save output to (eg '/home/puma/Desktop/Parts/')}

\item{sampsize}{Percent of animals to use to train model (default percent = 80)}

\item{prepboots}{bootstraps to use when preparing variables/model (~20 is sufficient)}

\item{finboots}{bootstraps to use when creating final model (200-1000)}

\item{max_fix}{max fix rate to test expressed as number (ie 12)}
}
\value{
outputs a new folder for each fix rate, goes from 1 to max_fix
}
\description{
This function tests the effects of rarefying your parturition model on model accuracy. You
can use this to help determine the fix rate needed for you data/study/species
}
\examples{
\donttest{fixrate(input = "/home/puma/Desktop/RunningPart/Data/StarkeyElk_ReadyData.RData", name = 'StarkElk', saveby = 'Study', ncpu = 75, basefolder = '/home/puma/Desktop/RunningPart/NewRarefy/', sampsize = 80, prepboots = 75, finboots = 100, max_fix = 12)}

}
\keyword{fix}
\keyword{fixrate}
\keyword{parturition}
\keyword{rate,}
