% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_modTrain.R
\name{Part_modTrain}
\alias{Part_modTrain}
\title{Wrapper function to create all variables necessary to create ungulate parturition
prediction models.}
\usage{
Part_modTrain(nrun, data, vars, mt, sampsize, idname = "UAID",
  partname = "birth", timename = "time",
  partdoyname = "Date.of.Birth")
}
\arguments{
\item{nrun}{number of bootstraps to run #vector('list',200)}

\item{data}{prep[[1]][[1]]}

\item{vars}{output of varprep}

\item{mt}{output of mt}

\item{sampsize}{percent of sample to use, 80 default}

\item{idname}{column name of unique animal id}

\item{partname}{column name of birth column}

\item{timename}{column name of posix time}

\item{partdoyname}{column name of julian day column}
}
\value{
Original data with rolling candidate variables
}
\description{
Wrapper function to create all variables necessary to create ungulate parturition
prediction models.
}
\keyword{part,}
\keyword{parturition}
