% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_mt.R
\name{Part_mt}
\alias{Part_mt}
\title{Wrapper function to create all variables necessary to create ungulate parturition
prediction models.}
\usage{
Part_mt(x, vars)
}
\arguments{
\item{x}{[[1]][[1]] from modPrep output}

\item{vars}{output of varPrep}
}
\value{
Original data with rolling candidate variables
}
\description{
Wrapper function to create all variables necessary to create ungulate parturition
prediction models.
}
\keyword{part,}
\keyword{parturition}
