% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_nPred.R
\name{Part_nPred}
\alias{Part_nPred}
\title{Wrapper function to create all variables necessary to create ungulate parturition
prediction models.}
\usage{
Part_nPred(rfmod, folder, datatype, mean_date, idname = "UAID",
  timename = "time")
}
\arguments{
\item{rfmod}{output of varprep}

\item{folder}{number of bootstraps to run #vector('list',200)}

\item{datatype}{prep[[1]][[1]]}

\item{mean_date}{output of mt}

\item{idname}{percent of sample to use, 80 default}

\item{timename}{column name of unique animal id}
}
\value{
Original data with rolling candidate variables
}
\description{
Wrapper function to create all variables necessary to create ungulate parturition
prediction models.
}
\keyword{part,}
\keyword{parturition}
