% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_npredResults.R
\name{Part_npredResults}
\alias{Part_npredResults}
\title{Wrapper function to create all variables necessary to create ungulate parturition
prediction models.}
\usage{
Part_npredResults(x)
}
\arguments{
\item{x}{output of nPred}
}
\value{
Original data with rolling candidate variables
}
\description{
Wrapper function to create all variables necessary to create ungulate parturition
prediction models.
}
\keyword{part,}
\keyword{parturition}
