% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_predResults.R
\name{Part_predResults}
\alias{Part_predResults}
\title{Wrapper function to create all variables necessary to create ungulate parturition
prediction models.}
\usage{
Part_predResults(x)
}
\arguments{
\item{x}{output of modTrain}
}
\value{
Original data with rolling candidate variables
}
\description{
Wrapper function to create all variables necessary to create ungulate parturition
prediction models.
}
\keyword{part,}
\keyword{parturition}
