% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_PredWrap.R
\name{Part_PredWrap}
\alias{Part_PredWrap}
\title{Wrapper function to create all variables necessary to create ungulate parturition
prediction models.}
\usage{
Part_PredWrap(rfmod, dat, mean_date, idname = "UAID",
  timename = "time", projstring, dataset, time.zone, ncpus, folder,
  type = "NPred", imp = FALSE)
}
\arguments{
\item{dat}{cleaned movement data as data.frame}

\item{mean_date}{Either 'Animal' or 'Study'. Most should use 'Study'}

\item{ncpus}{proj4string of dat projection}

\item{folder}{character vector shortname of dataset (eg. 'SMElk')}
}
\value{
List of length 2. 1) aggregates the model training - prediction results, 2) aggregates the predicted results to make a final prediction. Also saves all steps in /Results folder within the folder given to function
}
\description{
Wrapper function to create all variables necessary to create ungulate parturition
prediction models.
}
\keyword{part,}
\keyword{parturition}
\keyword{parturition,}
\keyword{prediction}
