% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_rollfun.R
\name{Part_rollfun}
\alias{Part_rollfun}
\title{Parturition Rolling Variable Creation}
\usage{
Part_rollfun(sub, original = FALSE)
}
\arguments{
\item{sub}{data}

\item{original}{Experimental, leave as default (FALSE)}
}
\value{
Original data with rolling candidate variables
}
\description{
Creates rolling candidate variables for ungulate partition prediction.
Internally used in parturition methodology.
}
\keyword{part,}
\keyword{parturition}
