% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_trajfun.R
\name{Part_trajfun}
\alias{Part_trajfun}
\title{Parturition Trajectory Creation}
\usage{
Part_trajfun(data, timename, xname, yname, idname)
}
\arguments{
\item{data}{"Cleaned" data}

\item{timename}{column name for posix time}

\item{xname}{column name for x coordinates}

\item{yname}{column name for y coordinates}

\item{idname}{column name for unique individual identifier}
}
\value{
Resulting object is the original data with trajectory variables attached
}
\description{
Creates trajectory variables for ungulate partition prediction.
Internally used in parturition methodology.
}
\keyword{part,}
\keyword{parturition}
