% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_varPrep.R
\name{Part_varPrep}
\alias{Part_varPrep}
\title{Wrapper function to create all variables necessary to create ungulate parturition
prediction models.}
\usage{
Part_varPrep(prep)
}
\arguments{
\item{prep}{output of modprep}
}
\value{
Original data with rolling candidate variables
}
\description{
Wrapper function to create all variables necessary to create ungulate parturition
prediction models.
}
\keyword{part,}
\keyword{parturition}
