% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartTab.R
\name{PartTab}
\alias{PartTab}
\title{Create Tables for Parturition Markdown}
\usage{
PartTab(vhist, mlist, vi, viout, outtra, spp)
}
\arguments{
\item{vhist}{Lookup table}

\item{mlist}{Character vector of mortality serials}

\item{vi}{Collar Download Status data}

\item{viout}{VIT Location data}

\item{outtra}{movement data}

\item{spp}{Species working with}
}
\value{
Returns plots for markdown in specified folder
}
\description{
Use movement and VIT data to create tables for Parturition Markdown Document
}
\examples{
\donttest{tabby<-tabFun(vhist=vhist,mlist=mlist,vi=vi,viout=viout,outtra=mdat2,spp='deer')}

}
\keyword{VIT,}
\keyword{markdown}
\keyword{parturition,}
