% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part_Wrap.R
\name{PartWrap}
\alias{PartWrap}
\title{Wrapper function to create all variables necessary to create ungulate parturition
prediction models.}
\usage{
PartWrap(dat, projstring, dataset, saveby, time.zone, ncpus, folder,
  mean_date, bday_dat, idname = "UAID", sampsize = 80, prepBoots,
  finBoots, imp = FALSE)
}
\arguments{
\item{dat}{cleaned movement data as data.frame}

\item{projstring}{proj4string of dat projection}

\item{dataset}{character vector shortname of dataset (eg. 'SMElk')}

\item{saveby}{Either 'Animal' or 'Study'. Most should use 'Study'}

\item{time.zone}{posix compliant time.zone character vector ('Etc/GMT-8')}

\item{ncpus}{number of cpus to use for parallel processing (minimum 2)}

\item{folder}{directory to store data in. If not created, function will create}

\item{mean_date}{julian date (as.numeric) of the mean date of parturition for your animal/system.}

\item{bday_dat}{data.frame with UAID and Date.of.Birth columns only}

\item{idname}{Column name (character vector) for Unique animal ID column (default = 'UAID')}

\item{sampsize}{Percent of animals to use to train model (default percent = 80)}

\item{prepBoots}{bootstraps to use when preparing variables/model (~20 is sufficient)}

\item{finBoots}{bootstraps to use when creating final model (200-1000)}
}
\value{
List of length 2. 1) aggregates the model training - prediction results, 2) aggregates the predicted results to make a final prediction. Also saves all steps in /Results folder within the folder given to function
}
\description{
Wrapper function to create all variables necessary to create ungulate parturition
prediction models.
}
\keyword{part,}
\keyword{parturition}
\keyword{parturition,}
\keyword{prediction}
