% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyData.R
\name{PrettyData}
\alias{PrettyData}
\title{Make data pretty}
\usage{
PrettyData(dat, idl, filen = "C:/Users/mhayes1/Desktop/fawnStuff/L12.rds")
}
\arguments{
\item{dat}{movement data}

\item{idl}{output of PartTab function}

\item{filen}{path to save rds file}
}
\value{
Pretty data for markdown file
}
\description{
Merge movement data and last fix information into pretty data for markdown
}
\examples{
\donttest{L12(dat=mdat2,idl=tabby,filen='/home/mhayes1/Desktop/DEERPTesting/L12.rds')}

}
\keyword{markdown}
