% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendUpdate.R
\name{sendUpdate}
\alias{sendUpdate}
\title{Email Auto Update}
\usage{
sendUpdate(from, to, subject = "Test Email from R",
  SP = "smtp.office365.com:587", attachpath,
  progpath = "C:/Users/mhayes1/sendEmail/sendEmail.exe", username,
  password, systype = "Win")
}
\arguments{
\item{from}{email address sending from}

\item{to}{email address sending to}

\item{subject}{email subject}

\item{SP}{server and port of outgoing email address}

\item{attachpath}{full path to attachment for email}

\item{progpath}{path to sendEmail program}

\item{username}{username for your email login}

\item{password}{password for your email login}
}
\description{
After the parturition update has been created, this function automatically sends the update
}
\examples{
\donttest{sendUpdate(from="mhayes1@uwyo.edu",to='mhayes1@uwyo.edu',
                     subject='Test email',SP="smtp.office365.com:587",
                     attachpath='C:/Users/mhayes1/Desktop/attach.xls',
                     progpath='C:/Users/mhayes1/sendEmail/sendEmail.exe',
                     username='mhayes1@uwyo.edu',password='password')}

}
\keyword{email,}
\keyword{update}
