% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SirTrackDat.R
\name{SirTrackDat}
\alias{SirTrackDat}
\title{Download data from Sirtrack Website}
\usage{
SirTrackDat(user, pass, saveas, yourlink)
}
\arguments{
\item{user}{username used to login to website}

\item{pass}{password used to login to website}

\item{saveas}{name of file (as .csv) to save data locally}

\item{yourlink}{this is the link used to download your data. Login to your sirtrack project,
hover over the "Export Full Data Set" csv icon, right click, Copy link address, and paste
as a character vector}
}
\value{
Resulting object is a spatial points data frame
\strong{Spatial Data Column Description:}
\item{CollarSerialNumber}{Collar Serial Number}
\item{TelemDate}{POSIX Field of collare fixes in the USER'S timezone. NOTE: may need to alter timezone}
\item{HDOP}{Horizontal Dilution of Precision}
\item{2D/3D}{Whether fix is a 2d or 3d fix. Values are either 2 or 3}
\item{Temperature}{Temperature reported by collar}
}
\description{
Download all data from a Sirtrack project, save as a csv locally and return a
spatialpointsdataframe of the data
}
\examples{
\donttest{stdat<-SirTrackDat(user=email,
pass=password,saveas='C:/Users/mhayes1/Desktop/SirtrackDat.csv')}

}
\keyword{sirtrack}
