% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajfun.R
\name{trajfun}
\alias{trajfun}
\title{Movement Trajectory}
\usage{
trajfun(data, timename, xname, yname, idname)
}
\arguments{
\item{data}{data.frame containing GPS data}

\item{timename}{name of POSIX field}

\item{xname}{name of x coordinate field}

\item{yname}{name of y coordinate field}

\item{idname}{name of individual identifier field}
}
\description{
Use adehabitatLT to create trajectory information for movement data and append the columns
to input spatial data. Internally called function; should not be used on it's own.
}
