% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecVitDat.R
\name{vecVitDat}
\alias{vecVitDat}
\title{Organize Vectronic Vit Data}
\usage{
vecVitDat(path)
}
\arguments{
\item{path}{path to vectronic data folder}
}
\value{
Resulting object is a data.frame of vectronic VIT data
}
\description{
Access and organize Vectronic VIT Data to use in daily Fawn Updates
}
\examples{
\donttest{vecVit<-vecVitDat(path='F:/Box Sync/DEER/Data/Vectronic/VecData')}

}
\keyword{vectronic,}
\keyword{vit}
