% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitMap.R
\name{vitMap}
\alias{vitMap}
\title{Create Plots for Parturition Markdown}
\usage{
vitMap(locdat, vidat, vhist, fold, spp, plotdataPath, hg = NULL)
}
\arguments{
\item{locdat}{movement data}

\item{vidat}{VIT data}

\item{vhist}{Lookup table}

\item{fold}{Folder to save plots in}

\item{spp}{Species working with}

\item{plotdataPath}{path ('C:/Users/mhayes1/Desktop/PlotData') to save plot data}
}
\value{
Returns plots for markdown in specified folder
}
\description{
Use movement data to create plots for Parturition Markdown Document
}
\examples{
\donttest{system.time({ vitMap(locdat=mdat2,vidat=vi,vhist=vhist,fold='/home/mhayes1/Desktop/DEERPTesting/plots',spp='deer') })
)}

}
\keyword{VIT,}
\keyword{markdown}
\keyword{parturition,}
